same modification for photon MVA
